requirements.txt
# RoyYoutubeSearch + RoyStudio 통합 의존성
# Python 3.9+ 필요

# ========== Web UI Framework ==========
eel>=0.16.0

# ========== Google API (YouTube Search) ==========
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# ========== Google Cloud TTS (RoyStudio) ==========
google-cloud-texttospeech>=2.14.0

# ========== 비디오/오디오 처리 (RoyStudio) ==========
moviepy>=2.0.0
pillow>=10.0.0
pydub>=0.25.1
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.1
matplotlib>=3.7.0

# ========== 텍스트 처리 ==========
python-docx>=1.1.0
kss>=4.0.0

# ========== 음성 인식 ==========
openai-whisper>=20231117

# ========== 네트워크/유틸리티 ==========
feedparser>=6.0.10
aiohttp>=3.9.0
cryptography>=42.0.0
openpyxl>=3.1.0
requests>=2.31.0
decorator>=5.1.1

# ========== PDF 처리 ==========
PyPDF2>=3.0.0

# 참고: FFmpeg는 별도로 설치하고 PATH에 추가해야 합니다
# 다운로드: https://ffmpeg.org/download.html
